feat(iluvatar): add Iluvatar Corex MR-V100 backend for KsanaLLM
Complete Iluvatar GPU support for KsanaLLM inference engine, enabling
TinyLlama-1.1B (and LLaMA-family models) to run end-to-end on MR-V100.

Key components:
- Build system: WITH_ILUVATAR cmake option, iluvatar.cmake toolchain
- Device layer: iluvatar_context/device (CUDA-compatible via Corex SDK)
- Kernel layer: bf16 RMSNorm, SwiGLU, RoPE, Embedding, Residual
- Attention: cuinfer FMHA for prefill, paged_attention for decode
- Sampling: cuinferTopKBatchV2 replacing naive single-thread argmax (42x faster)
- 3rdparty/LLM_kernels: iluvatar paged_attention, embedding, rotary_embedding

Benchmark (TinyLlama-1.1B, MR-V100 single GPU, 220 prompts):
- Token throughput: 2549 tokens/s
- vs NVIDIA A10: 2.16x gap (down from 2.46x after argmax optimization)

Includes quickstart.sh, QUICKSTART.iluvatar.md, and example config tuning.

+++ src/ksana_llm/models/gpt/gpt_config.h
@@ -68,7 +68,7 @@ inline void PrepareGPT2Attributes(const nlohmann::json &config_json, ModelConfig
   model_config.inter_size = config_json.value("n_inner", 4 * model_config.hidden_units);
   model_config.layernorm_eps = config_json.value("layer_norm_epsilon", 1e-5);
   model_config.start_id = config_json.value("bos_token_id", 50256);
-  model_config.end_ids = std::vector<uint32_t>{config_json.value("eos_token_id", 50256)};
+  model_config.end_ids = std::vector<uint32_t>{static_cast<uint32_t>(config_json.value("eos_token_id", 50256))};
   model_config.pad_id = config_json.value("pad_token_id", 0);
   model_config.max_position_embeddings = config_json.value("n_positions", 1024);
   model_config.activation_function = config_json.value("activation_function", "gelu_new");
